+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   * GPIO now relies on `getGPIO()` to determine board compatibility. It returns `None`
     if no library is available to access GPIO on this board.
 
+* **Fixed**
+  * MCP230xx cleanup pin configured log entries.
+
 ## 0.8.0
 
 * **Added**